Generate rxXYZ method with Single now
