+++ rx-java/src/main/asciidoc/dataobjects.adoc
@@ -873,34 +873,9 @@ Set whether client auth is required
 +++
 Set whether client auth is required
 +++
-|[[compressionLevel]]`compressionLevel`|`Number (int)`|
-+++
-This method allows to set the compression level to be used in http1.x/2 response bodies 
- when compression support is turned on (@see setCompressionSupported) and the client advertises
- to support <code>deflate/gzip</code> compression in the <code>Accept-Encoding</code> header
- 
- default value is : 6 (Netty legacy)
- 
- The compression level determines how much the data is compressed on a scale from 1 to 9,
- where '9' is trying to achieve the maximum compression ratio while '1' instead is giving
- priority to speed instead of compression ratio using some algorithm optimizations and skipping 
- pedantic loops that usually gives just little improvements
- 
- While one can think that best value is always the maximum compression ratio, 
- there's a trade-off to consider: the most compressed level requires the most
- computatinal work to compress/decompress data, e.g. more dictionary lookups and loops.
- 
- E.g. you have it set fairly high on a high-volume website, you may experience performance degradation 
- and latency on resource serving due to CPU overload, and, however - as the comptational work is required also client side 
- while decompressing - setting an higher compression level can result in an overall higher page load time
- especially nowadays when many clients are handled mobile devices with a low CPU profile.
- 
- see also: http://www.gzip.org/algorithm.txt
-+++
 |[[compressionSupported]]`compressionSupported`|`Boolean`|
 +++
-Set whether the server should support gzip/deflate compression 
- (serving compressed responses to clients advertising support for them with Accept-Encoding header)
+Set whether the server supports compression
 +++
 |[[crlPaths]]`crlPaths`|`Array of String`|
 +++
@@ -909,10 +884,6 @@ Add a CRL path
 |[[crlValues]]`crlValues`|`Array of Buffer`|
 +++
 Add a CRL value
-+++
-|[[decompressionSupported]]`decompressionSupported`|`Boolean`|
-+++
-Set whether the server supports decompression
 +++
 |[[enabledCipherSuites]]`enabledCipherSuites`|`Array of String`|
 +++
