+++ rx-java/src/main/resources/vertx-rxjava/template/class.templ
@@ -23,6 +23,7 @@ package @{type.raw.translatePackageName("rxjava")};\n\n
 
 import java.util.Map;\n
 import rx.Observable;\n
+import rx.Single;\n
 @comment{"Generate the imports"}
 @foreach{importedType:importedTypes}
 	@if{importedType.kind == CLASS_API}
