+++ rx-java/src/test/java/io/vertx/rx/java/test/JavaIntegrationTest.java
@@ -15,6 +15,7 @@ import io.vertx.rxjava.core.eventbus.MessageConsumer;
 import io.vertx.rxjava.core.http.HttpClient;
 import io.vertx.rxjava.core.http.HttpClientRequest;
 import io.vertx.rxjava.core.http.HttpClientResponse;
+import io.vertx.rxjava.core.http.HttpResponse;
 import io.vertx.rxjava.core.http.HttpServer;
 import io.vertx.rxjava.core.http.HttpServerRequest;
 import io.vertx.rxjava.core.http.ServerWebSocket;
@@ -28,6 +29,7 @@ import io.vertx.test.core.VertxTestBase;
 import org.junit.Test;
 import rx.Observable;
 import rx.Observer;
+import rx.Single;
 import rx.Subscriber;
 
 import java.util.ArrayList;
@@ -526,6 +528,31 @@ public class JavaIntegrationTest extends VertxTestBase {
           testComplete();
         });
       }).end();
+    });
+    await();
+  }
+
+  @Test
+  public void testHttpClientRequetBuilder() {
+    int times = 5;
+    waitFor(times);
+    HttpServer server = vertx.createHttpServer(new HttpServerOptions().setPort(8080));
+    server.requestStream().handler(req -> {
+      req.response().setChunked(true).end("some_content");
+    });
+    server.listen(ar -> {
+      HttpClient client = vertx.createHttpClient(new HttpClientOptions());
+      Single<HttpResponse<Buffer>> single = client
+        .createGet(8080, "localhost", "/the_uri")
+        .bufferBody()
+        .rxSend();
+      for (int i = 0;i < times;i++) {
+        single.subscribe(resp -> {
+          server.close();
+          assertEquals("some_content", resp.body().toString("UTF-8"));
+          complete();
+        }, this::fail);
+      }
     });
     await();
   }
